backend/requirements.txt
mysql.connector
flask_cors
flask_dropzone
flask_jwt_extended
bcrypt
